requirements.txt
gradio
python-dotenv
azure-ai-inference
azure-core
